move ctcs violation functions back into violation
